Refactor encodings into their own folder.
This will better split out the logic and make it easier to add decoders
in https://github.com/Yelp/py_zipkin/pull/100
