+++ tests/encoding/_encoders_test.py
@@ -3,7 +3,11 @@ import socket
 import mock
 import pytest
 
-from py_zipkin import _encoding_helpers
+from py_zipkin.encoding._helpers import create_endpoint
+from py_zipkin.encoding._encoders import get_encoder
+from py_zipkin.encoding._encoders import IEncoder
+from py_zipkin.encoding._encoders import _V1ThriftEncoder
+from py_zipkin.encoding._encoders import _V1JSONEncoder
 from py_zipkin import Encoding
 from py_zipkin.exception import ZipkinError
 
@@ -11,7 +15,7 @@ from py_zipkin.exception import ZipkinError
 @mock.patch('socket.gethostbyname', autospec=True)
 def test_create_endpoint_defaults_service_name(gethostbyname):
     gethostbyname.return_value = '0.0.0.0'
-    endpoint = _encoding_helpers.create_endpoint(port=8080)
+    endpoint = create_endpoint(port=8080)
 
     assert endpoint.service_name == 'unknown'
     assert endpoint.port == 8080
@@ -22,7 +26,7 @@ def test_create_endpoint_defaults_service_name(gethostbyname):
 @mock.patch('socket.gethostbyname', autospec=True)
 def test_create_endpoint_correct_host_ip(gethostbyname):
     gethostbyname.return_value = '1.2.3.4'
-    endpoint = _encoding_helpers.create_endpoint(host='0.0.0.0')
+    endpoint = create_endpoint(host='0.0.0.0')
 
     assert endpoint.service_name == 'unknown'
     assert endpoint.port == 0
@@ -34,7 +38,7 @@ def test_create_endpoint_correct_host_ip(gethostbyname):
 def test_create_endpoint_defaults_localhost(gethostbyname):
     gethostbyname.side_effect = socket.gaierror
 
-    endpoint = _encoding_helpers.create_endpoint(
+    endpoint = create_endpoint(
         port=8080,
         service_name='foo',
     )
@@ -45,7 +49,7 @@ def test_create_endpoint_defaults_localhost(gethostbyname):
 
 
 def test_create_endpoint_ipv6():
-    endpoint = _encoding_helpers.create_endpoint(
+    endpoint = create_endpoint(
         port=8080,
         service_name='foo',
         host='2001:0db8:85a3:0000:0000:8a2e:0370:7334',
@@ -57,7 +61,7 @@ def test_create_endpoint_ipv6():
 
 
 def test_malformed_host():
-    endpoint = _encoding_helpers.create_endpoint(
+    endpoint = create_endpoint(
         port=8080,
         service_name='foo',
         host='80',
@@ -70,19 +74,19 @@ def test_malformed_host():
 
 def test_encoder():
     assert isinstance(
-        _encoding_helpers.get_encoder(Encoding.V1_THRIFT),
-        _encoding_helpers._V1ThriftEncoder,
+        get_encoder(Encoding.V1_THRIFT),
+        _V1ThriftEncoder,
     )
     assert isinstance(
-        _encoding_helpers.get_encoder(Encoding.V1_JSON),
-        _encoding_helpers._V1JSONEncoder,
+        get_encoder(Encoding.V1_JSON),
+        _V1JSONEncoder,
     )
     with pytest.raises(ZipkinError):
-        _encoding_helpers.get_encoder(None)
+        get_encoder(None)
 
 
 def test_iencoder_throws_not_implemented_errors():
-    encoder = _encoding_helpers.IEncoder()
+    encoder = IEncoder()
     with pytest.raises(NotImplementedError):
         encoder.fits(0, 0, 0, "")
     with pytest.raises(NotImplementedError):
@@ -95,7 +99,7 @@ class TestBaseJSONEncoder(object):
     @pytest.fixture
     def encoder(self):
         """Test encoder"""
-        return _encoding_helpers.get_encoder(Encoding.V1_JSON)
+        return get_encoder(Encoding.V1_JSON)
 
     def test_fits(self, encoder):
         # count=2, current_size = 30, max_size = 52, len(new_span) = 20
@@ -106,7 +110,7 @@ class TestBaseJSONEncoder(object):
         assert encoder.fits(2, 30, 56, '{"trace_id": "1234"}') is True
 
     def test_create_json_endpoint(self, encoder):
-        ipv4_endpoint = _encoding_helpers.create_endpoint(
+        ipv4_endpoint = create_endpoint(
             port=8888,
             service_name='test_service',
             host='127.0.0.1',
@@ -117,7 +121,7 @@ class TestBaseJSONEncoder(object):
             'ipv4': '127.0.0.1',
         }
 
-        ipv6_endpoint = _encoding_helpers.create_endpoint(
+        ipv6_endpoint = create_endpoint(
             port=8888,
             service_name='test_service',
             host='2001:0db8:85a3:0000:0000:8a2e:0370:7334',
@@ -128,7 +132,7 @@ class TestBaseJSONEncoder(object):
             'ipv6': '2001:0db8:85a3:0000:0000:8a2e:0370:7334',
         }
 
-        v1_endpoint = _encoding_helpers.create_endpoint(
+        v1_endpoint = create_endpoint(
             port=8888,
             service_name=None,
             host='2001:0db8:85a3:0000:0000:8a2e:0370:7334',
@@ -139,7 +143,7 @@ class TestBaseJSONEncoder(object):
             'ipv6': '2001:0db8:85a3:0000:0000:8a2e:0370:7334',
         }
 
-        v2_endpoint = _encoding_helpers.create_endpoint(
+        v2_endpoint = create_endpoint(
             port=8888,
             service_name=None,
             host='2001:0db8:85a3:0000:0000:8a2e:0370:7334',
